Added role base function

+++ database/migrations/2024_11_26_132756_create_tbl_core_customer_table.php
@@ -11,20 +11,25 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('tbl_core_customer', function (Blueprint $table) {
-            $table->id();
-            $table->string('user_id'); // References 'users' table
-            $table->string('firstname');
-            $table->string('lastname');
-            $table->string('address');
-            $table->string('gender');
-            $table->string('email')->unique(); // Unique email
-            $table->string('phone_number');
-            $table->string('user_images')->nullable();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');                // Define the foreign key constraint
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('tbl_core_customer')) { // Check if the table exists
+            Schema::create('tbl_core_customer', function (Blueprint $table) {
+                $table->id();
+                $table->string('user_id'); // References 'users' table
+                $table->string('firstname');
+                $table->string('lastname');
+                $table->string('address');
+                $table->string('gender');
+                $table->string('email')->unique(); // Unique email
+                $table->string('phone_number');
+                $table->string('user_images')->nullable();
+                $table->timestamps();
+
+                // Define the foreign key constraint
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
+
 
     /**
      * Reverse the migrations.
